Don't special-case FLUSH
Previously, we'd do FLUSH from a separate thread, not affiliated with
the thread pool. While this works, it makes making sure there are no
outstanding flush requests at shutdown time much harder.

Therefore, change it so that the FLUSH command is handled by the (new)
thread pool, too; this way, we don't lose track of it.