Sanitize
Previously we pushed all queue handling to the backend, which would mean
that all backends would have to implement such queue handling
themselves.

This is silly, if not supid.

This changes the planned API so that a backend implements only the "push
data to the client" and/or "read data from the client" parts of the
system, and nbd-server itself does queue handling and/or retrial of
requests on block